Add ability to save state and start the status server
The state saving works like saving the pidfile, but it contains much
more rich information about the running server.

Additionally, added the --status option to start the status server so
that the server can be controlled over HTTP.

The state file contains the information about where the status server is
located.