Allows the server runloop to be optionally executed outside of a thread.
This patch unobtrusively fixes Server#run so that it can be executed
in the main thread. This is primarily to allow users to wrap run in
whatever thread spawning mechanism they'd like to, and enables use
with Celluloid by including the Celluloid module at load time:

  Puma::Server.send(:include, Celluloid)

allowing supervision of the server's runloop.