Add options to daemonize puma
Previously I was opposed to these options because I felt like puma
should always be run under another monitoring process. But after doing
some thinking (and some deployment of puma) I realized that the daemon
option is simply too useful to not include.

I highly suggest it be used with -w, so that if a worker crashes the
whole server doesn't mysteriously disappear.