Add preload_app!
This feature tells puma, when in cluster mode, to load the application
inside the master before booting the workers. This means that puma is
read to serve requests when it binds to the ports (important for listen
detection routers like Heroku). It has the side benefit of allowing very
fast worker boots.

Sadly though, this feature conflicts with the phased restart feature.
Because the application is loaded into the master process, there is no
clean way to boot new workers into a new application. As such, that
feature is disabled when preload_app! is used.

This gives you, the puma user, the ability to trade off features. You
can either restart workers quickly OR you can gracefully upgrade your
running application BUT you can't do both. Life is full of compromises.