pass raw IO objects to IO.select in reactor
Maglev currently has an issue where IO.select is not coercing its arguments to IO objects so passing Puma::Client objects to IO.select was raising an error.
This commit changes the reactor loop where we call IO.select so that it passes only IO objects to IO.select and then matches them back to the Puma::Client object.