cluster: add worker_boot_timeout option
When your workers are loading, the normal operation timeout is in
effect. If you spawn many workers or have them perform a lot of work
or wait for just long enough before actually booting, they will be
killed by the worker ping timeout.

If you want to control the exact timeout for your workers to boot,
specify the `worker_boot_timeout` option. Otherwise, it defaults to
the value of the `worker_timeout` option.