Set a more explicit SERVER_SOFTWARE Rack variable
When you need to have a switch in an infrastructure
middleware (streaming, events and sockets) you
might have to make runtime decisions based
on the server that is being used. The Rack env
header is the only way for the underlying Rack app
to see which server it is being run on. Previously
it would only contain the puma version string, which
is insufficient.