Add http logging to parse error
Sometimes I get errors like the one described in
https://github.com/puma/puma/issues/650. I made this PR with the
intention of printing the malformed request to STDOUT (at least for
MRI). I hope this accomplishes that.

I think this is a better alt to the error message and I don't think
there's a downside figuring the event is (hopefully) rare and log
management can easily handle these messages nowadays.