Prevent accumulation of clients being polled forever.
When a file descriptor for a client is returned from select, it
is removed from the clients currently being monitored for a
timeout. This removal happens independently from the code that
removes the descriptor from the list of descriptors that are
currently being polled, allowing a potential accumulation of
descriptors being passed into by select.