Apply default Host/Port to default_options
And not the default_config, which is a `Puma::DSL` object.

This aims to work in the spirit of the original code, but doesn't allow
the default_options inadvertently apply two different listening bind
addresses.

What this change fixes:
-----------------------
When the Rack::Handler::Puma sets the Puma::Configuration, it checks to
see if any :user_supplied_options exist in the options hash, and if that
hash exists, it then takes any keys not existing in that hash, and
pushes them down as default_options, giving them lesser priority when
evaluated.

This works fine for most values, except the `:binds`, since that can
sometimes be the combination of `:Host`/`:Port`.  When the Configuration
block is hit in the Rack handler, the `:binds` option already exists
with a default from `Puma::Configuration.puma_default_options`, so the
resulting config DSL will now append to that array, instead of creating
a new value.

This is most likely not the desired behavior, as this will create two
listening ports by default.

What this change does:
----------------------
This change effectively creates a `binds` key in the default_options,
based off of the `:Host` and `:Port` values, which will then override
the `Puma::Configuration.puma_default_options`'s `binds` key when the
`Puma::Configuration` instance is instanciated from the rack handler.

To do this, the `set_host_port_to_config` has been updated to work with
either a Puma::DSL (previous implementation), or a hash.  This allows us
to call it prior to instanciating the `Puma::Configuration`, instead of
during the configure block, preventing two listening ports from being
defined.