Appropriately wait for worker child process when shutting down via SIGTERM
This way when a SIGTERM is issued to a puma cluster (master process)
the shutting down of workers/child processes successfully completes.

Before this, it meant the master process will wait for every child process, not just
puma specific child process. Which meant even though the workers were to
successfully terminate, the server never exits as intended.