`ps -p` will check correctly whether $PID exists
It is appropriate to use `ps -p` in that place.

Under certain conditions, existing code ``[ "`ps -A -o pid=
| grep -c $PID`" -eq 0 ]`` may behave incorrectly.

For example, if the content of $PID happens to be a short string
(such as `199`), and another process includes the string
(process such as `2199`), then unfortunately `grep -c` will return
positive number even if `199` is not running.