Go back to using Bundler.with_original_env
Test that the environment is not lost when doing a restart by using the
PATH env var as a canary. While we actually care about the
BUNDLER_GEMFILE env var, that one is harder to set up in testing since
it requires a full copy of Puma's Gemfile.

This change also has the benefit of not only protecting the
BUNDLE_GEMFILE var, but also the other variables that Bundler knows
about.