[ci skip] Require Changelog at PR time
We frequently forget to ask for a changelog entry, and then when we want to release the release maintainer has to go back and manually add them all. It would be better if the changelog was required at PR time so the original submitter can think of the best wording. If a PR is trivial people can add `[ci skip]` or `[changelog skip]`.

This implementation uses github actions.