Implement separate tokens for control and status actions
Prior to this commit ...

Authentication for control and status actions is implemented via the same token.

With this commit ...

Puma implements a 'status-token' limited to status actions in Puma::App::Status.

If 'control-token' is defined, it is required for any action.
If 'control-token' is undefined, no token is required for any action.
If 'status-token' is undefined, no status token is required for status actions.

Closes #2081