Fix Connection header when not using queue_requests
When running with queue_requests=false, make the Connection header
reflect that the server will close the connection after the request is
completed.

(This is required by
https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html, specifically:
"HTTP/1.1 applications that do not support persistent connections MUST
include the "close" connection option in every message.")

This could previously cause clients which looked at this header to
attempt to reuse connections which the server had closed.