Create worker processes with `spawn` instead of `fork`...
...in clusters that support phased restarts.

If worker processes are created from the puma master process with just a
`fork`, then any gems loaded by the puma master are already "activated"
in the workers. This is mitigated in part through the use of the
`prune_bundler` configuration option, which ejects the puma master from
the application bundler context on startup, unloading all gems except
those used by the puma master process (typically, that's nio4r as well
as anything specified in `extra_runtime_dependencies`).

A fairly typical release strategy demonstrates the incompleteness of
`prune_bundler`. Imagine that in order to release a new version of an
application, users add the new version to the server, change a release
symlink to point the new version, initiate a phased restart on the
cluster, and then delete any old version of the application from disk
(including the gems). When a worker starts up and executes
`Bundler.setup`, it'll try to verify that each gem with native
extensions has been compiled by inspecting the gem directory on disk.
Since workers necessarily share the gems activated by the puma master,
Bundler will look for the gems on disk asscociated with the release that
was active at the time the puma master process started. If those gems
have since been deleted, Bundler raises `GemNotFound` and the workers
fail to boot.

This happens most frequently with the nio4r gem since it is 1) loaded
by the puma master process (even when `prune_bundler` is enabled), and
2) it has native extensions. But the same can happen to any gem that
matches those two conditions, such as the `ffi` gem if the puma master
is configured to use `puma_worker_killer`.

The fix introduced in this commit creates workers using `spawn`
instead of `fork`. `spawn` creates a new subprocess with a separate
memory space (`spawn` uses `fork` then `exec` on Unix-like
systems). The result is that workers share no gems with the puma master.
Workers can therefore start up successfully even if the master's gems
have been removed from disk. It's also possible to use different
versions of gems in the workers than what's loaded into the master.

The cost is some additional memory usage since workers that fork from
the master process can't leverage COW memory optimization. However, the
old `fork`-based method is still used in two cases:

1) If `preload_app` is enabled, since if `preload_app` is enabled,
   phased restarts aren't supported anyway. This option behaves as it
   did before: it forgoes the ability to perform phased restarts but
   maximizes the opportunity to use COW memory.
2) When workers fork from other workers (when `fork_worker` is enabled).
   Similar to `preload_app`, the main benefit of this option is to
   reduce memory by leveraging COW memory. The new spawn-based strategy
   is used only when creating `worker 0`, the process from which other
   workers fork.