Update extra_runtime_dependencies test to test master proccess's gems
The original motivation for `extra_runtime_dependencies` was to
activate gems in the puma master process after "pruning" the master
process with `prune_bundler`. This is useful for activating gems that
need to be loaded in the master process, such as `puma_worker_killer`.

The change to use `spawn` to create new worker processes does _not_
break this: `extra_runtime_dependencies` still works in that the gem
will still be activated as expected in the puma master process. However,
a side-effect of `extra_runtime_dependencies` was that those
dependencies would be loaded in puma worker processes as well (since
workers were just forked from the master).

The integration test was testing that `extra_runtime_dependencies` were
loaded in the worker processes instead of in the master process. This
change modifies the test to make assertions on the master processes'
gems directly.