Escape SSL cert and filenames
Filenames with `+` would fail to load because it would be
misinterpreted as a space. Since Puma already unescapes the query
string, we should escape the value to fix this.

We also make `cert` and `key` conditional in the query string because
that gives a clearer error message:

```
ERROR: Please specify the SSL key via 'key=' or 'key_pem='
```

Rather than:

```
puma/lib/puma/minissl.rb:218:in `check_file': Key file '' does not exist (ArgumentError)
```